feat(subscription): convert subscription page into a modal
